requirements.txt
python-dotenv
bs4
requests
huggingface_hub
beautifulsoup4
python-dotenv
lxml
requests
torch
pymongo
textwrap3
python-slugify
schedule
newspaper3k
slugify
lxml_html_clean
sacremoses
selenium
webdriver-manager